+++ app/scripts/mount_app.py
@@ -10,6 +10,7 @@ from src.hid_device import HID
 from src.msc_device import MSC
 from src.rndis_device import RNDIS
 from src.ncm_device import NCM
+from src.mtp_device import MTP
 from src.usb_peripheral import USBPeripheral
 
 paramdict = literal_eval(sys.argv[1])
@@ -102,6 +103,19 @@ class DeviceOperator():
             return True
         return False
     
+    def _isMTP(self) -> bool:
+        if self.simulator_action.split(' ')[0] == "MTP":
+            str_PIDVID = self.simulator_action.split(' ')[1:]
+            self.device_desc.idProduct = str_PIDVID[-1]
+            self.device_desc.idVendor = str_PIDVID[-2]
+            self.device_desc.bDeviceClass = 0x06
+            self.device_desc.bDeviceSubClass = 0x01
+            self.device_desc.bDeviceProtocol = 0x01
+            self.device_desc.product = "Emulated MTP device"
+            self.device_desc.bmAttributes = 0x80
+            return True
+        return False
+    
     def _isEJECT(self) -> bool:
         if self.simulator_action == "EJECT":
             return True
@@ -133,6 +147,7 @@ class DeviceOperator():
         ECM
         CDC
         NCM
+        MTP
         '''
         if self._isMSC():
             self.device_dict.fill_msc_dictionary()
@@ -160,6 +175,9 @@ class DeviceOperator():
 
         elif self._isNCM():
             self.device.usb_device = NCM(self.device_desc, DeviceFunction.ncm.value)
+
+        elif self._isMTP():
+            self.device.usb_device = MTP(self.device_desc, DeviceFunction.mtp.value)
 
         elif self._isEJECT():
             self._eject_device()